requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.4.2
aioredis==2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
tenacity==8.2.3
prometheus-client==0.19.0 